requirements.txt
flask
flask_cors
flask_sqlalchemy
flask_login
flask_httpauth
sqlalchemy_serializer
flask-bcrypt 
Flask-JWT-Extended
redis